chore: Move features to own internal package
* Card ID: CCT-1662
* Moved features.go to ./internal/features/
* Modified usage of package accordingly
* No functional changes
